Use correct path in vmtools
The ISO will end up in the home directory of the SSH user (unless
`tools_upload_path` is explicitly set, which it isn't here); since
this user name is an input, we shouldn't assume that it is "vagrant".

This was fixed as a side effect of #8, but broken again when that was
reverted in commit 0a7b1475.