Rename "ubuntu-server" task -> "server"
The Ubuntu config was using tasksel to add the "ubuntu-server" task, which
doesn't exist. Maybe it used to. The "server" task is actually what we want,
and pulls in packages for a headless setup, such as tmux.

The "ubuntu-server" task was essentially a no-op, so it didn't break anything,
but it did cause a lot of packages to be omitted. See full comparison
here https://gist.github.com/conorsch/7123209a86a2c577ba465b80cb05b0ef